# docs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: DPO оператора
  description: "REST API центра управления согласиями субъектов персональных данны\
    х. Идентификаторы — на английском, пользовательские тексты — на русском (NFR-8)."
  license:
    name: Proprietary
  title: In consensu — центр управления согласиями
  version: v1
servers:
- description: Текущий хост
  url: /
paths:
...
  /api/v1/consent-types/{code}/deactivate:
    post:
      operationId: deactivate_3
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConsentTypeResponse"
          description: OK
      tags:
      - consent-type-controller
...
  /api/v1/forms/{id}/history:
    get:
      operationId: history_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryResponse"
          description: OK
      tags:
      - consent-form-controller
...
  /api/v1/notification-rules/{id}/deactivate:
    post:
      operationId: deactivate_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotificationRuleResponse"
          description: OK
      tags:
      - notification-rule-controller
...
  /api/v1/subjects:
    get:
      operationId: search
      parameters:
      - in: query
        name: query
        required: false
        schema:
          type: string
      - in: query
        name: pageable
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseSubjectResponse"
          description: OK
      tags:
      - subject-controller
    post:
      operationId: upsert
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertSubjectRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubjectResponse"
          description: OK
      tags:
      - subject-controller
  /api/v1/subjects/by-external-id/{externalId}:
    get:
      operationId: getByExternalId
      parameters:
      - in: path
        name: externalId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubjectResponse"
          description: OK
      tags:
      - subject-controller
  /api/v1/subjects/by-external-id/{externalId}/card:
    get:
      operationId: cardByExternalId
      parameters:
      - in: path
        name: externalId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubjectCardResponse"
          description: OK
      tags:
      - subject-controller
  /api/v1/subjects/{id}:
    get:
      operationId: get_7
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubjectResponse"
          description: OK
      tags:
      - subject-controller
  /api/v1/subjects/{id}/card:
    get:
      operationId: card
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubjectCardResponse"
          description: OK
      tags:
      - subject-controller
  /api/v1/subjects/{id}/card.pdf:
    get:
      operationId: cardPdf
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/pdf:
              schema:
                type: string
                format: byte
          description: OK
      tags:
      - subject-controller
...
  /api/v1/subjects/{id}/history:
    get:
      operationId: history
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsentResponse"
          description: OK
      tags:
      - subject-controller
...
  /api/v1/third-parties:
    get:
      operationId: list_2
      parameters:
      - in: query
        name: pageable
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseThirdPartyResponse"
          description: OK
      tags:
      - third-party-controller
    post:
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThirdPartyRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ThirdPartyResponse"
          description: Created
      tags:
      - third-party-controller
  /api/v1/third-parties/{id}:
    get:
      operationId: get_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ThirdPartyResponse"
          description: OK
      tags:
      - third-party-controller
    put:
      operationId: update_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThirdPartyRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ThirdPartyResponse"
          description: OK
      tags:
      - third-party-controller
  /api/v1/third-parties/{id}/deactivate:
    post:
      operationId: deactivate_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ThirdPartyResponse"
          description: OK
      tags:
      - third-party-controller
...
  /api/v1/webhooks/{id}/deactivate:
    post:
      operationId: deactivate
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WebhookSubscriptionResponse"
          description: OK
      tags:
      - webhook-controller
...
components:
  schemas:
...
    Basis:
      type: object
      properties:
        consentId:
          type: string
          format: uuid
        typeCode:
          type: string
        validUntil:
          type: string
          format: date-time
...
    ChannelView:
      type: object
      properties:
        allowed:
          type: boolean
        basis:
          $ref: "#/components/schemas/Basis"
        channel:
          type: string
        channelRu:
          type: string
        reason:
          type: string
        reasonRu:
          type: string
...
    ConsentResponse:
      type: object
      properties:
        daysLeft:
          type: integer
          format: int64
        form:
          $ref: "#/components/schemas/FormRef"
        grantedAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        pdnCategories:
          type: array
          items:
            type: string
        purposes:
          type: array
          items:
            type: string
        revocationSource:
          type: string
        revokedAt:
          type: string
          format: date-time
        signatureType:
          type: string
        source:
          type: string
        sourceRef:
          type: string
        status:
          type: string
        statusText:
          type: string
        subjectId:
          type: string
          format: uuid
        thirdPartyId:
          type: string
          format: uuid
        thirdPartyName:
          type: string
        typeCode:
          type: string
        typeId:
          type: string
          format: uuid
        typeNameRu:
          type: string
        validUntil:
          type: string
          format: date-time
...
    ConsentTypeResponse:
      type: object
      properties:
        active:
          type: boolean
        businessSignificant:
          type: boolean
        category:
          type: string
        categoryRu:
          type: string
        channels:
          type: array
          items:
            type: string
        code:
          type: string
        defaultValidity:
          type: string
        dependsOnCode:
          type: string
        description:
          type: string
        id:
          type: string
          format: uuid
        nameRu:
          type: string
        requiresThirdParty:
          type: boolean
        sortOrder:
          type: integer
          format: int32
...
    ContactRequest:
      type: object
      properties:
        primary:
          type: boolean
        type:
          type: string
          enum:
          - PHONE
          - EMAIL
          - POSTAL_ADDRESS
        value:
          type: string
          maxLength: 512
          minLength: 0
      required:
      - type
      - value
    ContactResponse:
      type: object
      properties:
        masked:
          type: boolean
        primary:
          type: boolean
        type:
          type: string
        typeRu:
          type: string
        value:
          type: string
...
    CreateThirdPartyRequest:
      type: object
      properties:
        inn:
          type: string
          pattern: "\\d{10}|\\d{12}"
        thirdParty:
          $ref: "#/components/schemas/ThirdPartyRequest"
      required:
      - inn
      - thirdParty
...
    FormRef:
      type: object
      properties:
        checksum:
          type: string
        code:
          type: string
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int32
...
    HistoryResponse:
      type: object
      properties:
        actor:
          type: string
        at:
          type: string
          format: date-time
        comment:
          type: string
        decision:
          type: string
        eventType:
          type: string
        eventTypeRu:
          type: string
        fromStatus:
          type: string
        role:
          type: string
        toStatus:
          type: string
...
    NotificationRuleResponse:
      type: object
      properties:
        active:
          type: boolean
        channels:
          type: array
          items:
            type: string
        consentTypeId:
          type: string
          format: uuid
        daysBefore:
          type: array
          items:
            type: integer
            format: int32
        id:
          type: string
          format: uuid
        name:
          type: string
        recipientEmails:
          type: array
          items:
            type: string
        recipientRoles:
          type: array
          items:
            type: string
        thirdPartyId:
          type: string
          format: uuid
        triggerType:
          type: string
        triggerTypeRu:
          type: string
...
    PageResponseSubjectResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/SubjectResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageResponseThirdPartyResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ThirdPartyResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    SubjectCardResponse:
      type: object
      properties:
        channels:
          type: array
          items:
            $ref: "#/components/schemas/ChannelView"
        channelsSummaryRu:
          type: string
        consents:
          type: array
          items:
            $ref: "#/components/schemas/ConsentResponse"
        generatedAt:
          type: string
          format: date-time
        subject:
          $ref: "#/components/schemas/SubjectResponse"
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/TransferView"
...
    SubjectResponse:
      type: object
      properties:
        birthDate:
          type: string
          format: date
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/ContactResponse"
        externalId:
          type: string
        firstName:
          type: string
        fullName:
          type: string
        id:
          type: string
          format: uuid
        lastName:
          type: string
        middleName:
          type: string
...
    ThirdPartyRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
    ThirdPartyRequest:
      type: object
      properties:
        address:
          type: string
        allowedPdnCategories:
          type: array
          items:
            type: string
          uniqueItems: true
        contactEmail:
          type: string
          maxLength: 255
          minLength: 0
        contractDate:
          type: string
          format: date
        contractNumber:
          type: string
          maxLength: 128
          minLength: 0
        contractValidUntil:
          type: string
          format: date
        name:
          type: string
          maxLength: 512
          minLength: 0
        ogrn:
          type: string
          maxLength: 15
          minLength: 0
        role:
          type: string
          enum:
          - PROCESSOR
          - RECIPIENT
          - ECOSYSTEM
        shortName:
          type: string
          maxLength: 255
          minLength: 0
      required:
      - address
      - name
      - role
    ThirdPartyResponse:
      type: object
      properties:
        active:
          type: boolean
        address:
          type: string
        allowedPdnCategories:
          type: array
          items:
            type: string
        contactEmail:
          type: string
        contractDate:
          type: string
          format: date
        contractDaysLeft:
          type: integer
          format: int64
        contractExpired:
          type: boolean
        contractNumber:
          type: string
        contractValidUntil:
          type: string
          format: date
        id:
          type: string
          format: uuid
        inn:
          type: string
        name:
          type: string
        ogrn:
          type: string
        role:
          type: string
        roleRu:
          type: string
        shortName:
          type: string
...
    TransferView:
      type: object
      properties:
        basisConsentId:
          type: string
          format: uuid
        daysLeft:
          type: integer
          format: int64
        pdnCategories:
          type: array
          items:
            type: string
        thirdParty:
          $ref: "#/components/schemas/ThirdPartyRef"
        thirdPartyContractExpired:
          type: boolean
        validUntil:
          type: string
          format: date-time
...
    UpsertSubjectRequest:
      type: object
      properties:
        birthDate:
          type: string
          format: date
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/ContactRequest"
        externalId:
          type: string
          maxLength: 128
          minLength: 0
        firstName:
          type: string
          maxLength: 128
          minLength: 0
        lastName:
          type: string
          maxLength: 128
          minLength: 0
        middleName:
          type: string
          maxLength: 128
          minLength: 0
      required:
      - externalId
      - firstName
      - lastName
...
    WebhookSubscriptionResponse:
      type: object
      properties:
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        eventTypes:
          type: array
          items:
            type: string
        id:
          type: string
          format: uuid
        name:
          type: string
        updatedAt:
          type: string
          format: date-time
        url:
          type: string